requirements.txt
et_xmlfile==2.0.0
greenlet==3.2.4
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
playwright==1.54.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
